+++ flask_express/response.py
@@ -102,7 +102,7 @@ class Response(ResponseBase):
         if isinstance(content, (dict, list)):
             return self.json(content)
 
-        if isinstance(content, Munch): # for munch object.
+        if isinstance(content, Munch) or issubclass(content, Munch): # for munch object.
             return self.json(content.toDict())
 
         try:
@@ -112,7 +112,7 @@ class Response(ResponseBase):
         except JSONDecodeError:
             return self.make_response(content)
 
-        except Exception as e:
+        except Exception:
             raise TypeError("provided data type is not supported for `send` function.")
 
     def json(self, *wargs:t.Any, **kwargs:t.Any) -> "Response":
@@ -526,18 +526,26 @@ class Response(ResponseBase):
                 headers=self.headers
                 )
 
-    def response_from_obj(self, rv:"Response") -> "Response":
+    def make_response_from_obj(self, rv:"Response", headers=None, status=None) -> "Response":
         """
         take a response object as the parameter and 
         return the new response object from the old instance.
 
         :param rv: the response object
         """
-        return self.make_response(
-            rv.response,
-            rv.status,
-            rv.headers,
-            rv.mimetype,
-            rv.content_type,
-            rv.direct_passthrough,
-        )
+        self.status = rv.status
+        self.status_code = rv.status_code
+        self.headers = rv.headers
+        self.content_type = rv.content_type
+        self.direct_passthrough = rv.direct_passthrough
+
+        if headers is not None:
+            self.headers.update(headers)
+        
+        if status is not None:
+            if isinstance(status, (str, bytes, bytearray)):
+                self.status = status  # type: ignore
+            else:
+                self.status_code = status
+        
+        return self.make_response(rv.response)