+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 0.1.0
 
-- Renamed the module name to flask-express from flaske to make it a extension of Flask.
+- renamed the module name to flask-express from flaske to make it a extension of Flask.
 
 ## 0.1.1
 
 - Some typo fixed at README.md.
 - Some typo fixed at Munch module.
+- Aded the proper logo.